Closes #5 revision table

+++ database/migrations/2020_07_07_084632_create_sponsorships_table.php
@@ -16,8 +16,7 @@ class CreateSponsorshipsTable extends Migration
         Schema::create('sponsorships', function (Blueprint $table) {
             $table->id();
             $table->float('price', 3 , 2);
-            $table->unsignedSmallInteger('duration');
-            //$table->timestamps();
+            $table->unsignedTinyInteger('duration');
         });
     }
